Add application configuration and setup http initializer
