+++ frontend/package.json
@@ -5,6 +5,8 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "build:dev": "ng build --configuration=development",
+    "build:prod": "ng build --configuration=production",
     "watch": "ng build --watch --configuration development",
     "format": "prettier --config .prettierrc --write \"src/**/*.{ts,js,html,json,scss}\"",
     "test": "jest",
