+++ frontend/src/app/core/services/config.service.ts
@@ -13,7 +13,7 @@ export class ConfigService {
 
   // Charge le fichier JSON correspondant à l'environnement
   loadConfig(env: string): Promise<void> {
-    const configPath = `assets/config/environment.${env}.json`;
+    const configPath = `./environments/app-config.${env}.json`;
     return this.http.get(configPath).toPromise().then((config) => {
       this.config = config;
     });
