--- frontend/src/assets/config/environment.dev.json
@@ -1,4 +1,0 @@
-{
-    "apiBaseUrl": "http://localhost:3000/",
-    "production": false
-  }
