--- frontend/src/assets/config/environment.prod.json
@@ -1,4 +1,0 @@
-{
-    "apiBaseUrl": "TODO",
-    "production": true
-  }
