+++ frontend/src/environments/app-config.development.json
@@ -1,0 +1,3 @@
+{
+    "apiBaseUrl": "http://localhost:3000"
+  }
