+++ frontend/src/environments/app-config.production.json
@@ -1,0 +1,3 @@
+{
+  "apiBaseUrl": "TODO"
+}
