+++ frontend/src/environments/environment.development.ts
@@ -1,0 +1,4 @@
+export const environment = {
+    name: 'development',
+    production: false,
+};
