+++ frontend/src/environments/environment.production.ts
@@ -1,0 +1,4 @@
+export const environment = {
+    name: 'production',
+    production: true,
+};
