+++ frontend/src/environments/environment.ts
@@ -1,0 +1,7 @@
+// Recommandation from https://v17.angular.io/guide/build
+
+export const environment = {
+    name: '',
+    production: false,
+};
+
